refactor(core): extract maintenance/narrative-extract.ts — pure body-text parser
