+++ packages/core/src/maintenance.ts
@@ -1833,23 +1833,9 @@ type DedupKeyCandidateRow = {
  * sections found in session_summary body_text. Returns null if the body
  * doesn't match the expected structure.
  */
-export function extractNarrativeFromBody(bodyText: string): string | null {
-	// Match sections like "## Completed\n...\n\n## Learned\n..."
-	const sections: string[] = [];
+export { extractNarrativeFromBody } from "./maintenance/narrative-extract.js";
 
-	const completedMatch = bodyText.match(/##\s*Completed\s*\n([\s\S]*?)(?=\n##\s|\n*$)/);
-	if (completedMatch?.[1]?.trim()) {
-		sections.push(completedMatch[1].trim());
-	}
-
-	const learnedMatch = bodyText.match(/##\s*Learned\s*\n([\s\S]*?)(?=\n##\s|\n*$)/);
-	if (learnedMatch?.[1]?.trim()) {
-		sections.push(learnedMatch[1].trim());
-	}
-
-	if (sections.length === 0) return null;
-	return sections.join("\n\n");
-}
+import { extractNarrativeFromBody } from "./maintenance/narrative-extract.js";
 
 /**
  * Backfill narrative for session_summary memories that have structured
